Moved ElkLogger functionality into its own module
This is to reduce confusion when maintaining Logger.py
